Committing this ,but it has an a Mocha error
Added imageanimation component.
