+++ app/index.html
@@ -101,6 +101,7 @@
     <!-- load components, these provide webcomponents (with corresponding templates in app/templates)-->
     <!-- mainly tested using integration tests -->
     <!-- these should be replaceable if someone wants to switch to a different webcomponent platform -->
+    <script src="scripts/components/imageanimation.js"></script>
     <script src="scripts/components/modelcreate.js"></script>
     <script src="scripts/components/modeldetails.js"></script>
     <script src="scripts/components/modellist.js"></script>
