+++ app/scripts/components/modeldetails.js
@@ -1,25 +1,21 @@
-/* global fetchModel, fetchLog, deleteModel, startModel, stopModel, router */
+/* global ImageAnimation, fetchModel, fetchLog, deleteModel, startModel, stopModel, router */
 var exports = (function () {
   "use strict";
 
   var ModelDetails = Vue.component("model-details", {
     template: "#template-model-details",
 
+    components: {
+      // <my-component> will only be available in Parent's template
+      "image-animation": ImageAnimation
+
+    },
+
     // Show the details of one model
     data: function() {
       console.log("loading model details for id", this.id);
 
       return {
-        // Current animation frame:
-        currentAnimationIndex: 0,
-
-        // timer id for animation.
-        timerAnimation: -1,
-
-        // Which imagelist are we currently watching?
-        currentAnimationKey: "delta_fringe_images",
-
-        isAnimating: false,
 
         model: {
         }
@@ -75,10 +71,10 @@ var exports = (function () {
           this.updateData(parseInt(val));
 
           // Stop any animation if we were animating.
-          this.stopImageFrame();
+          //this.stopImageFrame();
 
           // Reset index:
-          this.currentAnimationIndex = 0;
+          //this.currentAnimationIndex = 0;
         }
 
 
@@ -101,30 +97,6 @@ var exports = (function () {
           return this.model.logoutput;
         }
       },
-      animationIndex: {
-        cache: false,
-        get: function() {
-          return this.currentAnimationIndex;
-        }
-      },
-
-      animationFrame: {
-        cache: false,
-        get: function() {
-          var animationKey = this.currentAnimationKey;
-
-          if (animationKey.length > 0) {
-            var imgs = this.model.info[animationKey];
-
-            if (imgs !== undefined) {
-              return this.model.fileurl + imgs.location + imgs.images[this.currentAnimationIndex];
-            }
-          }
-
-          return "";
-
-        }
-      },
 
       // Returns true if the model is running. We might have to depend on some other variables
       // But for now we say that "processing" means running.
@@ -133,28 +105,6 @@ var exports = (function () {
         get: function() {
 
           return this.model && this.model.state === "PROCESSING";
-        }
-      },
-
-      isAnimating: {
-        cache: false,
-        get: function() {
-          return this.timerAnimation > 0;
-        }
-      },
-      hasFrames: {
-        cache: false,
-        get: function() {
-
-          var animationKey = this.currentAnimationKey;
-          var imgs = this.model.info[animationKey];
-
-          if (imgs !== undefined) {
-            return imgs.images.length > 0;
-          }
-
-          return 0;
-
         }
       }
     },
@@ -278,37 +228,38 @@ var exports = (function () {
           });
       },
 
-      changeMenuItem: function(event) {
+      // Not used anymore?
+      // changeMenuItem: function(event) {
 
-        var el = $(event.target);
+      //   var el = $(event.target);
 
-        // Hide all panels except for the target.
-        $(".collapse").hide();
+      //   // Hide all panels except for the target.
+      //   $(".collapse").hide();
 
-        // Get target:
-        var targetSelector = $(el).attr("data-target");
-        var target = $(targetSelector);
+      //   // Get target:
+      //   var targetSelector = $(el).attr("data-target");
+      //   var target = $(targetSelector);
 
-        target.show();
-
-
-        // If there is an animation property, we set this:
-        var targetAnimation = $(el).attr("data-animation");
-
-        if (targetAnimation !== undefined && targetAnimation.length > 0) {
-          this.currentAnimationKey = targetAnimation;
-          this.currentAnimationIndex = 0;
-          this.stopImageFrame();
-
-        } else {
-          this.currentAnimationKey = "";
-          this.currentAnimationIndex = 0;
-        }
+      //   target.show();
 
 
+      //   // If there is an animation property, we set this:
+      //   var targetAnimation = $(el).attr("data-animation");
 
-        event.stopPropagation();
-      },
+      //   if (targetAnimation !== undefined && targetAnimation.length > 0) {
+      //     this.currentAnimationKey = targetAnimation;
+      //     this.currentAnimationIndex = 0;
+      //     this.stopImageFrame();
+
+      //   } else {
+      //     this.currentAnimationKey = "";
+      //     this.currentAnimationIndex = 0;
+      //   }
+
+
+
+      //   event.stopPropagation();
+      // },
 
       // User wants to start a model. We just do not do anything now, as this needs to be implemented.
       startModel: function() {
@@ -343,81 +294,7 @@ var exports = (function () {
 
       },
 
-      // For animations:
-      previousImageFrame: function() {
-        // Check if an animation key has been set. If not, we bail out.
-        if (this.currentAnimationKey.length === 0) {
-          return;
-        }
 
-        // Does not exist?
-        if (this.model.info === undefined) {
-          return;
-        }
-
-
-        this.currentAnimationIndex--;
-
-        var imgs = this.model.info[this.currentAnimationKey];
-
-
-        // Probably wrap with active key.
-        if (this.currentAnimationIndex < 0) {
-          this.currentAnimationIndex = imgs.images.length - 1;
-        }
-      },
-
-      stopImageFrame: function() {
-        // Check if an animation key has been set. If not, we bail out.
-        if (this.currentAnimationKey.length === 0) {
-          return;
-        }
-
-
-        // Clear interval
-        if (this.timerAnimation !== -1) {
-          //  this.isAnimating =  false;
-          clearInterval(this.timerAnimation);
-
-          this.timerAnimation = -1;
-        }
-
-      },
-
-      playImageFrame: function() {
-        // Check if an animation key has been set. If not, we bail out.
-        if (this.currentAnimationKey.length === 0) {
-          return;
-        }
-
-        // Stop and start. (We do not want multiiple setintervals)
-        this.stopImageFrame();
-        this.timerAnimation = setInterval(this.nextImageFrame, 1000);
-
-      },
-
-      nextImageFrame: function() {
-        // Check if an animation key has been set. If not, we bail out.
-        if (this.currentAnimationKey.length === 0) {
-          return;
-        }
-
-        // Does not exist?
-        if (this.model.info === undefined) {
-          return;
-        }
-
-        this.currentAnimationIndex++;
-
-        var imgs = this.model.info[this.currentAnimationKey];
-
-        if (imgs !== undefined) {
-          // Probably wrap.
-          if (this.currentAnimationIndex >= imgs.images.length) {
-            this.currentAnimationIndex = 0;
-          }
-        }
-      },
 
       downloadOptionsChange: function() {
