+++ app/styles/main.scss
@@ -417,6 +417,10 @@ html, body {
 
 }
 
+.animation-frame
+{
+  max-width: 100%;
+}
 
 .model-information {
   padding: 4px 0 0 32px;
