+++ app/templates/finder-columns.html
@@ -30,6 +30,8 @@
       </div>
     </nav>
 
+
+
     <alert-message></alert-message>
 
     <div class="container-fluid full-height">
